pyproject.toml
[tool.poetry]
name = "enable-depandabot-auto-commit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "enable_depandabot_auto_commit"}]

[tool.poetry.dependencies]
python = "^3.10"
GitPython = "^3.1.37"
PyYAML = "^6.0.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
